Move fixtures to fixture file
