+++ conftest.py
@@ -2,4 +2,5 @@
 
 pytest_plugins = [
     u'ckanext.harvest.tests.fixtures',
+    u'ckanext.xroad_integration.tests.fixtures'
 ]
